Reorganize topology management tests
